added AbstractStorageAdapter for common code opaque: moved put-transaction code to base class
