+++ cadc-storage-adapter-fs/src/main/java/org/opencadc/inventory/storage/fs/OpaqueIterator.java
@@ -3,7 +3,7 @@
 *******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
 **************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
 *
-*  (c) 2022.                            (c) 2022.
+*  (c) 2025.                            (c) 2025.
 *  Government of Canada                 Gouvernement du Canada
 *  National Research Council            Conseil national de recherches
 *  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -195,6 +195,6 @@ class OpaqueIterator implements Iterator<StorageMetadata> {
     }
     
     private StorageMetadata createStorageMetadata(Path p) {
-        return OpaqueFileSystemStorageAdapter.createStorageMetadata(contentPath, p, includeRecoverable);
+        return OpaqueFileSystemStorageAdapter.createStorageMetadataImpl(contentPath, p, includeRecoverable);
     }
 }
